+++ desktop/js/scenario.js
@@ -154,7 +154,7 @@ $(function(){
         for (var group in scenarioGroupedList) {
           items = {}
           for (var i in scenarioGroupedList[group]) {
-            scName = scenarioGroupedList[group][i].groupObjectName.replace('['+group+']', '')
+            scName = scenarioGroupedList[group][i].humanName.replace('['+group+']', '')
             scId = scenarioGroupedList[group][i].id
             items[uniqId] = {'name': scName, 'id' : scId}
             uniqId ++
@@ -1417,11 +1417,7 @@ function printScenario(_id) {
         updateTooltips()
       }, 500)
 
-      var group = $('input[data-l1key="group"]').val()
-      if (group == '') group = '{{Aucun}}'
-      var object = $('select[data-l1key="object_id"] option:selected').text().trim()
-      var name = $('input[data-l1key="name"]').val()
-      $('#groupObjectName').html('['+group+']['+object+']['+name+']')
+      $('#humanNameTag').html(data.humanNameTag)
     }
   })
 }